startup fixes, listen, TCP, multiple clients
Fixed and added code for dealing with multiple clients using listen(TCP, '0.0.0.0', <port>)

Also fixed a few inline comments according to PEP 8 -- Style Guide for Python Code
"An inline comment is a comment on the same line as a statement. Inline comments should be separated by at least two spaces from the statement. They should start with a # and a single space."